// API_Client/client/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My First Nest API
 * The API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface User {
    'id': number;
    'firstName': string;
    'lastName': string;
    'email': string;
    'role': UserRoleEnum;
    'gender'?: UserGenderEnum;
    'age'?: number;
    'password': string;
    'createdAt': string;
}

export const UserRoleEnum = {
    Admin: 'admin',
    User: 'user',
} as const;

export type UserRoleEnum = typeof UserRoleEnum[keyof typeof UserRoleEnum];
export const UserGenderEnum = {
    Male: 'male',
    Female: 'female',
} as const;

export type UserGenderEnum = typeof UserGenderEnum[keyof typeof UserGenderEnum];